Remove lodash from Tabs component
